ai_agent_based_operation/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
identify==2.6.2
idna==3.10
iniconfig==2.0.0
instructor==1.6.4
Jinja2==3.1.4
jiter==0.6.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
nodeenv==1.9.1
numpy==2.1.3
openai==1.55.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
propcache==0.2.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
-e git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47#egg=swarm
tenacity==9.0.0
tqdm==4.67.0
typer==0.13.1
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
yarl==1.18.0
